backend/kafka_producer/requirements.txt
kafka-python-ng==2.2.3
python-dotenv==1.0.1
requests==2.31.0

